Add border top + include aggregate footer in table
