+++ packages/twenty-front/src/modules/object-record/record-table/record-table-row/components/RecordTableTr.tsx
@@ -6,6 +6,10 @@ const StyledTr = styled.tr<{ isDragging: boolean }>`
       ? `1px solid ${theme.border.color.medium}`
       : '1px solid transparent'};
   transition: border-left-color 0.2s ease-in-out;
+
+  &:nth-last-child(2) td {
+    border-bottom: none;
+  }
 `;
 
 export const RecordTableTr = StyledTr;
